extracts and fixes types
