+++ src/types/index.ts
@@ -1,1 +1,9 @@
 export type EthAddress = { address: string };
+
+export * from './contract';
+export * from './activity';
+export * from './contractMatch';
+export * from './execute';
+export * from './tokenAllocation';
+export * from './proposal';
+export * from './transaction';
